+++ resources/views/components/community/user-card.blade.php
